more variable type cleanups

+++ lib/ftp.c
@@ -18,7 +18,7 @@
  * This software is distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY
  * KIND, either express or implied.
  *
- * $Id: ftp.c,v 1.233 2004-03-03 13:17:44 bagder Exp $
+ * $Id: ftp.c,v 1.234 2004-03-03 13:24:14 bagder Exp $
  ***************************************************************************/
 
 #include "setup.h"
@@ -2044,7 +2044,7 @@ CURLcode Curl_ftp_nextconnect(struct connectdata *conn)
         char *bytes;
         bytes=strstr(buf, " bytes");
         if(bytes--) {
-          int in=bytes-buf;
+          long in=bytes-buf;
           /* this is a hint there is size information in there! ;-) */
           while(--in) {
             /* scan for the parenthesis and break there */
@@ -2412,7 +2412,7 @@ CURLcode Curl_ftpsendf(struct connectdata *conn,
 {
   ssize_t bytes_written;
   char s[256];
-  ssize_t write_len;
+  size_t write_len;
   char *sptr=s;
   CURLcode res = CURLE_OK;
 
@@ -2436,7 +2436,7 @@ CURLcode Curl_ftpsendf(struct connectdata *conn,
     if(conn->data->set.verbose)
       Curl_debug(conn->data, CURLINFO_HEADER_OUT, sptr, bytes_written);
 
-    if(bytes_written != write_len) {
+    if(bytes_written != (ssize_t)write_len) {
       write_len -= bytes_written;
       sptr += bytes_written;
     }
